More bob docs updates (#2288)
* Improving cold start docs

* Removing workflow example in favor of code.

* Adding issue `config.yml` (suggested by GitHub)

* Renaming source re: issue #2192

* Closes #404

* Spellcheck
