+++ docs/docs/.vuepress/config.js
@@ -96,6 +96,7 @@ module.exports = {
               "/code/go/",
               "/code/bash/",
               "/destinations/",
+              "/scheduling-future-tasks/",
             ],
           },
           {
