+++ docs/docs/workflows/events/cold-starts/README.md
@@ -4,8 +4,44 @@ If your workflow doesn't process an event for roughly 5 minutes, Pipedream turns
 
 If your workflow needs to process data in a time-sensitive manner (for example, if you're issuing an HTTP response), you can implement the following workaround to keep your workflow "warm". 
 
-- Create a scheduled workflow that runs roughly every 5 minutes, making an HTTP request to your HTTP-triggered workflow on the `/warm` path ([see example workflow](https://pipedream.com/@dylburger/warm-up-http-workflow-p_A2CQ9ne/edit)).
-- Then, in your original workflow, add a step at the top that ends the workflow early if it receives a request on this `/warm` path. You can set this path to be whatever you'd like — `/warm` is just an example. On normal requests, that step won't run and your workflow will proceed as normal ([see example workflow](https://pipedream.com/@dylburger/end-early-on-warming-requests-p_PACqYrW/edit)).
+### 1. Create a scheduled workflow that triggers your original workflow via HTTP request
+
+First, create a scheduled workflow that runs roughly every 5 minutes, making an HTTP request to your HTTP-triggered workflow on the `/warm` path.
+
+Here's a Node.js example:
+
+```javascript
+import got from "got"
+
+export default defineComponent({
+  async run({ steps, $ }) {
+    return (await got(`${params.url}/warm`)).body
+  },
+})
+```
+
+### 2. End your original workflow on warming requests
+
+Then, in your original workflow, add a step at the top that ends the workflow early if it receives a request on this `/warm` path. You can set this path to be whatever you'd like — `/warm` is just an example. On normal requests, that step won't run and your workflow will proceed as normal.
+
+Here's a Node.js example:
+
+```javascript
+export default defineComponent({
+  async run({ steps, $ }) {
+    const path = (new URL(event.url)).pathname
+    if (path === "/warm") {
+      $.respond({
+        status: 200,
+        body: {
+          warmed: true,
+        }
+      })
+      return $.flow.exit("Warming request, ending early")
+    }
+  },
+})
+```
 
 We're tracking the ability to keep a workflow permanently warm [here](https://github.com/PipedreamHQ/pipedream/issues/318). Feel free to follow that issue to receive updates.
